pyproject.toml
[project]
name = "ids-uncertainty-estimation"
version = "1.0.0"
description = "LLM Uncertainty Estimation for Intrusion Detection Systems"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "openai>=1.3.0",
    "agno>=1.8.0",
    "groq>=0.4.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0"
]

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["E501"]
line-length = 100

[tool.black]
line-length = 100
